Bug 1243847 - Part 1: Merge default engine scraping into plugin scraping script
This allows us to run ./scrape_plugins.py to do a full import of all
engines and defaults, instead of spreading the logic across multiple
scripts.
